add sample-data.sql

+++ exam/springjpa/src/main/java/vn/techmaster/springjpa/repository/CourseRepository.java
@@ -8,7 +8,6 @@ import vn.techmaster.springjpa.entity.Course;
 public interface CourseRepository extends JpaRepository<Course, Long> {
 
     // Viết Native Query để tính điểm trung bình một môn bất kỳ
-
     @Query(value = "select avg(score) from course c inner join student_course sc on c.id = sc.course_id where c.name = ?1 group by c.id", nativeQuery = true)
-    float calAverageScoreByCourse(String courseName);
+    double calAverageScoreByCourse(String courseName);
 }
