+++ exam/springjpa/src/main/java/vn/techmaster/springjpa/service/CourseService.java
@@ -1,5 +1,5 @@
 package vn.techmaster.springjpa.service;
 
 public interface CourseService {
-    float avgScoreByCourse(String coursename);
+    double avgScoreByCourse(String coursename);
 }
